+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `README.md` updated
 - `404.html` links bug fixed
 - Window resizing bug fixed
+- Recent plays history size changed to 25
 ## [0.5] - 2025-12-17
 ### Added
 - Watch time