+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6] - 2026-01-06
 ### Added
 - `CODE_OF_CONDUCT.md`
 - Skip button
@@ -61,7 +62,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Load subtitle from URL
 
 
-[Unreleased]: https://github.com/sepandhaghighi/neovid/compare/v0.5...main
+[Unreleased]: https://github.com/sepandhaghighi/neovid/compare/v0.6...main
+[0.6]: https://github.com/sepandhaghighi/neovid/compare/v0.5...v0.6
 [0.5]: https://github.com/sepandhaghighi/neovid/compare/v0.4...v0.5
 [0.4]: https://github.com/sepandhaghighi/neovid/compare/v0.3...v0.4
 [0.3]: https://github.com/sepandhaghighi/neovid/compare/v0.2...v0.3